requirements.txt
gradio
transformers
sentence-transformers
faiss-cpu
torch
